Fix exception on writing Last COM Port setting to ASCOM profile
Writing the "Last COM Port" value to the ASCOM profile of the focuser
throws an exception, because the device type of the profile has not been
set to "Focuser", so it is "Telescope" by default.

This means that if auto-detecting the COM port is off in the focuser
app, the connection always fails.

The device type was being set inside an "if" clause in the code, meaning
it was never being set in the non-auto-detecting case.